Kubernetes docs: Replace static tables with the auto-generated tables
I found out that some properties were not synced between the content in the Kubernetes guide and the config-all page (which uses the auto generated table).

Let's use the auto-generated tables also in the Kubernetes guide, so we avoid having to maintain these properties, plus also having to explain the inner basic types (as it's already part of the auto generated table).