Introduce gRPC post-processing
When using gRPC, there is a code generation phase. The generator (protoc) generates a bunch of classes used to implement and consume the services.
This commit introduces a post-processing phase to that generation so we can modify the generated source files.

The rationale is twofold:

1. stubs cannot be mocked because of the `final` modifier in the source.
2. the generated classes use `javax.annotation.Generated`, which blocks the migration to Jakarta.

The current post-processing addresses these two issues. Each feature can be disabled individually. The whole post-processing can also be disabled.