Fix missing UdsNettyChannelProvider<init> in native mode
PR #26047 updated gRPC to 1.47.0.
However, gRPC is also used from OpenTelemetry, and the execution paths are slightly different, and #26047 did not cover that path.
As a result, gRPC tried to create an instance from a class in which the constructor has been eliminated by gRPC (on purpose, we have ensured it gets eliminated), so fail to create the instance.

This PR extends the checks when creating the instance to check for class issues and missing constructors.