Properly register classes linked to `CustomResource` extensions.
The `findWatchedClasses` method was formerly only working properly if
the target class for which extensions or implementations were examined
was an interface. However, `CustomResource` is an abstract class. This
fixes this problem. This also adds the extension itself as part of the
watched classes as previously, only the spec and status classes were
registered.