Handle null containerInfo in dev UI page for Dev Services config
This page has always existed, and there has always been an item for
"Other Dev Services" where the container info is not available but some
config is.

However, it seems the dev UI page has never worked when there
actually was "Other Dev Services" since containerInfo is null and the
dev UI page encounters an NPE when rendering this item.

Since we're now making it more likely that there is an "Other Dev
Services" item, because of the Hibernate ORM config we're adding,
the NPE became much more visible.