Bump to Vert.x 4.3.1 and Netty 4.1.78
This new combination is tricky. It imposes to have brotli4j available in the classpath as, because of the new HTTP compression feature from vert.x, it does now have a hard dependency on brotli4j. While in JVM mode, everything is fine, native compilation fails if not there. We have tried a few things to avoid the dependency, but we are hitting https://github.com/oracle/graal/issues/4652.