Always produce JSON build output for native
Latest GraalVM master (release 22.3.0 in October) has new
options so as to produce a JSON file listing relevant build
output metrics in that JSON file. In addition, there are options
to get some timing information (e.g. total time to build an image)
and get that written to a JSON file as well.

This patch enables those options for GraalVM 22.3.0 and better
as they are fairly low overhead in terms of compute resources and
filesystem space, yet are useful for after the fact analysis of the
build and for CI.