Fix #24308: improve container image repository checks for some dev services
This commit improves container image name checks for these dev services:

- Apicurio Registry
- ActiveMQ Artemis (AMQP)
- RabbitMQ

These dev services used to verify that the image repository is equal
to a hard-coded string. This commit changes that to verify that
the image repository _ends with_ the same hard-coded string, which
allows for repository namespacing in private container registries.