Reorder Resteasy Reactive handlers so preClientSendHandler is always first
Presumably the `preClientSendHandler` was added to support observability/telemetry but it has not been working.

The `preClientSendHandler` was previosuly ordered _after_ the request filters but the telemetry is reported using a filter (e.g. `OpenTelemetryClientFilter`). Given that setup the  client’s route template was not set on the context until after the telemeetry was reported.

This simply rearranges the `preClientSendHandler` to run first, before any other handlers, including filters.