fix #26899: Prefer get getters over is getters in recorders
When both get getters and is getters exist, they can
replace the original selected getter in PropertyUtils.
This can creates inconsistencies, resulting in a field
not being set, causing a NPE later down the line that is
hard to diagnose.

To remove inconsistencies, if there is both a get method
and an is method, the get method is always preferred.
This introduces consistent behavior for recorders.

To accomplish this, a seperate map is used for isGetters in
PropertyUtils. Using a seperate map removes the need to have a seperate
loop for identifing the actual getter method. It also
make the logic more clear (first try "get" getter, then try "is"
getter).