Deprecate HibernateValidatorFactoryCustomizer in favor of ValidatorFactoryCustomizer
And document it.
While working on the documentation, I realized that we are referencing
ValidatorFactory everywhere so it makes no sense to call the customizer
with the Hibernate prefix.

For now, I deprecated the class, it will be dropped in the next minor version
as it was only in one micro and the usage is relatively narrow.