Testing: Add failure state in testing context
Sometimes it's necessary to know whether the tests failed (at test case level or even in the before stages). For example: if we use one resource via the `@QuarkusTestResource` annotation that is starting a service and we don't want to stop this service when the test fails.

At the moment, we are not propagating when a test fails.

This pull request addresses this issue by adding this information in:
- QuarkusTestContext (for QuarkusTest*Callbacks)
- Context (for QuarkusTestResources)

To be precise, I've added the following methods in the above contexts:

```java
/**
         * Get the test failure result if the test failed. When no test have failed, it will return null.
         *
         * @return the failure result that is thrown during either `BeforeAll`, `BeforeEach`, test method, `AfterAll` or
         *         `AfterEach` phases.
         */
        TestStatus getTestStatus();
```

In the future, we could add even more information like the number of tests that have failed and/or the test profiles.

Fix https://github.com/quarkusio/quarkus/issues/25749