Improve SmallRye Health extension's default scope for health checks
The SmallRye Health extension registers health check annotations
as additional bean defining annotations, but doesn't define a default
scope for them. Instead, the extension used to produce an annotation
transformation that adds `@Singleton` to every class or method annotated
with one of the health check annotations and not annotated with a scope
annotation or a stereotype with a scope annotation.

This commit removes that annotation transformation and instead defines
that the health checks annotations that are registered as BDAs have
a default scope of `@Singleton`. Less code, same outcome.