Fix treatment of recursive bean types
ArC used to pretty much not care about recursive bean types, because
they are very rare. However, per the CDI specification, `@Dependent`
beans may declare type parameters, and those can easily be recursive.

This commit fixes the bytecode generator that instantiates the set
of bean types. Bean types are, at runtime, represented using the common
`java.lang.reflect.Type` hierarchy, which represents recursive types
as an infinite path through the object graph of types. We clearly can't
generate a bytecode sequence to instantiate an infinite number of types,
but we can use a trick inspired by Jandex 3.0: add a notion of type
variable references.

For example, in `<T extends Comparable<T>>`, the first occurence of `T`
is represented by a `TypeVariableImpl`, which has a parameterized type
as its bound. The type argument of that parameterized type, which is
the second occurence of `T`, is represeted by `TypeVariableReferenceImpl`.
In the generated bytecode, the types are instantiated from inside, so
the reference is instantiated first, and when the containing type variable
is instantiated, the reference is patched to delegate to it.