Replace usage of NeverInline in jdbc-pgsql
The previous implementation relied on a graal compiler optimization to
never inline a method. By doing so it tricked native-image into not
including XML processing code when it's not used.

Instead, use a DomHelper-local bifunction field which stays null unless
the XML processing code is actually reachable. If it is, the bifunction
is being set by the reachability handler and, thus, achieves the same
result.

Closes #27791