Use a timer when awaiting Quartz shutdown
The Quartz scheduler shutdown seems like it
can block indefinitely if there is an exception
in a yet to be completed job.
To guard against this, we ensure that we only
wait a configurable amount of time for the scheduler
to shut down.

This is a breaking change because previously we waited
indefinitely.

Relates to #28114

Co-authored-by: Manyanda Chitimbo <manyanda.chitimbo@gmail.com>