Add skeleton tests for extension descriptor
I've also corrected (because the tests caught it, yay!) a missing brace on one of the "please don't complain" properties.

I haven't tested the bootstrap extension descriptor since it is deprecated, but we could copy things across fairly easily if we think it's worthwhile.

I also haven't tested everything the descriptor does, but now that we have a skeleton which drives the mojo, we can add new tests more easily.