Run the method scheduled with the Quartz scheduler on a duplicated context.
This is the follow-up of https://github.com/quarkusio/quarkus/pull/28255.
When running the scheduled method, the lack of duplicated context can lead to contextual data loss (loss only, no leak).
This commit makes sure that the method runs with a duplicated context.

However, this is a breaking change for users expecting the method to run on a Quartz thread.