Register Resource methods for reflection if ResourceInfo#getMethod is used in filters
ResourceInfo#getMethod is used in order to retrieve the Resource method
that is handles the HTTP request, and it obviously uses reflection.
Resource methods are not by default registered for reflection, but this
change introduces a way for Quarkus to read the bytecode of all
filter methods and checks whether any of them uses
the ResourceInfo#getMethod API and if so, forces the registration
for reflection of all Resource methods.

Fixes: #28182