Generate the attributes.adoc file during the build
Currently, we maintain a copy of the attributes.adoc in the website and
we always get it wrong when we add a new attribute.

This way, attributes.adoc will come directly from the release and will
be complete.
Attributes that can be overridden has been moved to
attributes-local.adoc. attributes-local.adoc won't be overwritten when
syncing.