Ensure that Ryuk does not get started when checking for Docker
Usually Ryuk containers started by TestContainers will be stopped,
but there are cases when after a build such containers are left
dangling.
With this change, we ensure that Ryuk is not started when all
we are doing is checking whether docker is available.

Fixes: #25852
(cherry picked from commit 6c8eae801079b023409ad09f82075b4196fd7eed)