Update supported-by information for extension descriptors
In support of #8021, process the supported-by attribute to find a supporter name, then create or update an attribute with the supporter name and the suffix of `-support` to with the status(es) for the extension supporter.  If there is an existing tag ending with `-support`, a `supported-by` tag will be created updated for the supporter(s).

Signed-off-by:Nathan Erwin <nathan.d.erwin@gmail.com>