Rest Data Panache: Correct Open API integration
Before these changes, Open API could not deduct the return type when using Response.class (in Rest Data Panache classic). 
Also, for Rest Data Panache reactive, the return object for the endpoint '/get' was wrongly saying that was a single entity when is an array of entities. 
And finally, for the update and delete endpoints, the result status code was wrong (for update should be 201 and for delete 204).
After these changes, all the above issues are fixed and covered using the existing test.
Fix https://github.com/quarkusio/quarkus/issues/28786