Fix compatibility with Redis 5
The output format of a few commands has changed between Redis 5 and 6.
This PR allows us to verify the compatibility with Redis 5, 6, and 7.

Note that it's the best effort, and we do not intend full compatibility with Redis 5, only the main use cases. Typically, Redis Stack is only supported with Redis 7.