Use the GraalVM 22.3 AlwaysInline annotation
This annotation has moved to a new package and we need to use the 22.3
version to make sure our code compiles.
As for using 22.2, it still works but the annotation is ignored, which
looks OK as the annotation is not used in places where it is strictly
necessary.
If people use the recommended GraalVM version, everything will be fine.
If not, things will mostly work anyway.

(cherry picked from commit 6c5700f8b04eccf5adfb11631cf690d2d71b669c)