Use non closeable output stream for writting the multipart entities
Before, we were using a ByteArrayOutputStream because some message writers are closing the original output stream when writing the entity (like JSON).
The problem is that this approach needs the double of space (keeping the data in the bytearray output stream and then copy the data to the original output stream).
With these changes, we are proxying the original output stream and preventing the message writers to not close the output stream (this was already being used in the JsonbMessageBodyWriter).
Fix https://github.com/quarkusio/quarkus/issues/28920

(cherry picked from commit 25722b721b23a10eaec46430a5e52f0b33e715e2)