Prevent null pointer in setAllHeaders
I didn't add any test to cover this change because it's easier to fix than to replicate the conditions to reproduce it in either the extension or an integration test than expected taking into account we have a reproducer.
Still let me know if you want it to cover it and maybe I can add wiremock in the integration-tests/rest-client-reactive. 
Fix https://github.com/quarkusio/quarkus/issues/28922