Automatically register nested Kotlin classes for reflection
Users have come to expect that @RegisterForReflection fixes
a lot of native image issues and use it a lot.
For Kotlin classes however, registering a class is often not
enough - the companion classes are need to be registered.
These companion classes are nested classes, so we automatically
set the property to true for Kotlin classes

Closes: #28167
(cherry picked from commit 62748638fd3d2c0424eb71ef2707cdb5028ffcc3)