Fix logging with panache in interfaces
The logging transformation used to add a `private static final` field
to each transformed class. This is invalid in case the transformed
class is in fact an interface, because the JVMS requires interface
fields to always be `public static final`.

With this commit, the logging transformation still adds `private` fields
to regular classes, but when the transformed class is an interface,
the generated field is `public`. This is not very nice, but working
code trumps aesthetically pleasing code.

(cherry picked from commit 572a79015bfad68aff94afff00de34f6a7bdfbae)