Prevent on-demand-indexing warnings for missing meta-annotations
When the bean archive index is created, additional beans are indexed
and annotations present directly on those classes are indexed as well.
Meta-annotations are not.

On the other hand, the Spring DI extension tries to obtain all meta-
annotations from each indexed annotation, transitively, so when it
stumbles upon a missing annotation class, the on-demand-indexing
`IndexWrapper` prints a warning before remembering that such class
doesn't exist.

With this commit, the bean archive index creation process indexes
all meta-annotations on additional beans, transitively, and if some
of them is missing, the `IndexWrapper` remembers that and can later
immediately return `null` without attempting to load the class.

This prevents a meaningless warning for missing annotation classes,
which is safe, because annotations in Java are designed to be
potentially missing.

(cherry picked from commit d2d1b10a3b5a00f045e6d9c4d93d6581aa9dda26)