Separate resourceClass into resourceName and resourceInterface
The changes in https://github.com/quarkusio/quarkus/pull/29212 introduced a bug that was reproduced by the Quarkus Quickstart (see https://github.com/quarkusio/quarkus-quickstarts/actions/runs/3475885826). 

The problem is that in REST DATA with Panache, we use the REST interface to declare the resource properties and the generated resource extends to this REST interface. However, in Spring Data Rest, we use the repository to gather the resource properties and hence the generated resource extends to this repository and as we have two beans for the repository (the right one and the generated resource), it fails. 

The fix is to separate the resourceInterface into resourceName (where the resource properties can be set) and the good resourceInterface (the one we'll use to extend the generated resource - in Spring Data REST, this is null because we don't have resource interfaces).