Update protoc to 3.21.9.
This update uses https://github.com/protocolbuffers/protobuf/releases/tag/v21.7, which contains a fix for https://github.com/protocolbuffers/protobuf/security/advisories/GHSA-h4h5-3hr4-j3g2.
Note that because the fix for that CVE was backported to 3.19, Quarkus is not affected, but better to stay up to date when we can.

Fix #29317