ArC: filter out class-retained or missing annotations
Sometimes, ArC used to call the `AnnotationLiteralProcessor` with
annotations that are either class-retained, or their classes are missing
(which is perfectly legal). It is a minority of callers though -- most
of them only call `AnnotationLiteralProcessor` with qualifiers or
interceptor bindings. Those must always be runtime-retained and their
classes must be present, and it is an error if they are not.

At the same time, we don't want to generate annotation literals for
class-retained annotations or for annotations whose classes are missing.

The `AnnotationLiteralProcessor` already checks that the annotation class
exists, so with this commit, it also checks that the annotation for which
the literal is generated is runtime-retained. If not, an exception is
thrown. This lets us know that there's a problem somewhere.

There are 2 places where ArC deals with arbitrary annotations and may need
to generate annotation literals for them; in both cases, the code has to
process all annotations present on an injection point. In these cases, it
is straightforward to filter out class-retained annotations and missing
annotation classes.

(cherry picked from commit 46268e3a8e2383b12258645d0435019b1761fb7c)