Fix Vert.x event bus codec registration
The `EventBusCodecProcessor` used to create a map from Jandex `Type`
of message to a `DotName` of the corresponding codec, and then expose
that map as a repeatable `MessageCodecBuildItem`. When creating each
build item, the code used to call `Type.toString()`, which is never
the right thing to do when obtaining a class name from Jandex `Type`.

With this commit, the map is from type `DotName` to codec `DotName`,
and the key in the map is obtained by `Type.name()`.