ArC: fix assignability check during build-time resolution
The `AssignabilityCheck.isAssignableFrom()` method used to search a Jandex
index for subtypes of first type to determine if it is assignable from
the second type. This doesn't work well with on-demand indexing; for example,
if we ask the index for all subtypes of `Comparable`, we don't get anything,
because JDK types are not in the index.

This commit reverses the `isAssignableFrom()` logic: instead of searching
the index for subtypes of the first type, it searches the index for supertypes
of the second. That works well with on-demand indexing and doesn't change
meaning.

This change alone is enough to fix build-time resolution that involves
recursive generic type. For example, a `@Dependent` producer method for
type `List<T>`, where `T extends Comparable<T>`, now correctly matches
when resolving an injection point of type `List<String>`.