Optimize CsrfHandler use of SecureRandom
Ideally the `secureRandom` field would be `final`, however
we can't use that because it leads to a Random being captured
in the native image heap.
So instead we take advantage of the fact that `secureRandom`
is only set once and only by the main thread, which means
that we can use `setRelease` / `getAcquire` semantics in
order to avoid paying the cost of volatile access.