Add further details about podman
This adds further details about the fix and the discussion
in https://github.com/quarkusio/quarkus/pull/29562.
It clearifies that alias is also fine instead of podman-docker,
just redirects the user to podman for install instructions (
they change over time, and more linux variant are shown there,
no need duplicate and maintain), and finally,
it shows a way of running podman as service without
systemd (relevant on wsl up until recently for example,
and users a most likely stuck with slightly older wsl's in a few places).

Signed-off-by: Josef Andersson <josef.andersson@gmail.com>