Updates to the Cassandra quickstart guide
This commit updates the Cassandra quickstart guide and
clarifies the steps required in order to use the
Cassandra driver Object Mapper framework in a Quarkus
application.

It also makes it simpler to understand which beans
are automatically produced by the Object Mapper.

Finally, it brings some minor enhancements to the
"Eager vs Lazy Initialization" section.