Reduce some object allocations during bytecode recording
The idea behind the PR is that in most cases of bytecode recording,
the recorders simply read their input and don't change it.
By leveraging this insight, we can eliminate the allocations of
lists and maps that are empty by using the empty versions
of these types.

To make Quarkus do this however, we introduce the `@ReadOnly`
which extension developers need to add to recorder parameters
in order to let Quarkus know that it is safe to apply this
optimization