Move log4j-api and add-condition for jaxb-annotations registrations
log4j-api is a dependency of elasticsearch-rest-high-level-client

`log4j-api` is not always present when using the
elasticsearch-rest-client-common extension as it's not a dependency of
it. Instead, it is a dependency of elasticsearch-rest-high-level-client:

```
+- io.quarkus:quarkus-elasticsearch-rest-high-level-client:jar:999-SNAPSHOT:compile
|  +- ...
|  +- org.jboss.logmanager:log4j2-jboss-logmanager:jar:1.1.1.Final:compile
|  |  \- org.apache.logging.log4j:log4j-api:jar:2.19.0:compile
```

`com.fasterxml.jackson.module.jaxb.JaxbAnnotationIntrospector` is not
always present when using the quarkus-jackson extension as it's not a
dependency of it. Instead, it is a dependency of
quarkus-resteasy-jackson:

```
+- io.quarkus:quarkus-resteasy-jackson:jar:999-SNAPSHOT:compile
|  +- ...
|  +- org.jboss.resteasy:resteasy-jackson2-provider:jar:4.7.7.Final:compile
|  |  +- ...
|  |  +- com.fasterxml.jackson.jaxrs:jackson-jaxrs-json-provider:jar:2.14.1:compile
|  |  |  +- com.fasterxml.jackson.jaxrs:jackson-jaxrs-base:jar:2.14.1:compile
|  |  |  \- com.fasterxml.jackson.module:jackson-module-jaxb-annotations:jar:2.14.1:compile
```

This results in failed attempts to register the class for reflection
when using the `quarkus-jackson` extension without
`jackson-module-jaxb-annotations` in the classpath.

Since we expect users to manually include the artifact in their
dependencies even when not using `quarkus-resteasy-jackson` we maintain
the registration in the `quarkus-jackson` extension but only perform it
when the artifact is present in the classpath.