Add (short) name for class loaders
This sets the name that appears in stack traces for each of our major class loaders.

Also, ensure that the contract for `ClassLoader#getName()` is correctly met: "This method is non-final for compatibility. If this method is overridden, this method must return the same name as specified when this class loader was instantiated."

Qbicc compilation uses class loader name as a key to help with reproducible compilation.