ArC: fix runtime representation of recursive bean types
This was already attempted once, but the previous fix only worked well
for simple recursive generic types. This commit fixes the runtime
representation even for mutually recursive generic types.

The fix uses the following observation: in a type, multiple occurences
of a type variable (as denoted by the type variable identifier) always
refers to the same type. Therefore, it is not necessary to maintain
a stack of type variables as they being processed. Instead, a simple
string-keyed map is enough to:

1. canonicalize all occurences of a type variable;
2. canonicalize all occurences of a type variable reference;
3. at the end, patch all type variable references to point to
   the type variables.

The canonicalization was not performed before, but is necessarsy now
(to correctly handle equality of multiple occurences of type variables),
plus it also slightly improves runtime footprint of the set of bean types.