Use the new Jandex API for reconstructing descriptors and signatures
The `AsmUtil` classes in ArC and RESTEasy Reactive are modified freely,
as they are assumed to be private to these projects. The `AsmUtil` class
in Quarkus core is modified in a backward compatible fashion, as this
class can be used outside of main Quarkus code. All its methods that deal
with descriptor/signature reconstruction are marked as deprecated,
with the corresponding Jandex methods as suggested replacements.