Propagate security annotations in the Spring Data Rest for all methods
Before these changes, the security annotations were only propagated to the generated resource if and only if the method was annotated with either the `@RestResource` or `@RepositoryRestResource` annotation. 

In contract, the pure REST Data with Panache does not have this limitation (propagate these annotations regardless the annotation present). 

I've also updated the documentation to mention this feature.
Fix https://github.com/quarkusio/quarkus/issues/30358