Add default implementation for REST Data interfaces
The linked issue is caused by Arc because the method interfaces are automatically proxied when there are security annotations. And the problem is that this method are not implemented.

However, we can't implement these methods since the generated resources take care also of the response codes. 

Fix https://github.com/quarkusio/quarkus/issues/30515