Improve JDBC driver error message in Agroal processor
This improves the error message when the DB Kind is not found.

```
Unable to find a JDBC driver corresponding to the database kind 'postgres' for the default datasource (available: 'postgresql'). Check if it's a typo, otherwise provide a suitable JDBC driver extension, define the
driver manually, or disable the JDBC datasource by adding 'quarkus.datasource.jdbc=false' to your configuration if you don't need it.
```

Based on feedback from https://github.com/quarkusio/quarkus/issues/30701#issuecomment-1410757424