Ignore interface/class without default constructs fields in SB config
This should behave similarly that when the setter is private. In this case, the field is ignored and we write a warning saying that a field is going to be ignored.

Fix https://github.com/quarkusio/quarkus/issues/30896

(cherry picked from commit 999289a59e9befff6222d90027dcf4c33c3e9075)